+++ client/src/Components/Header/Navbar/Navbar.js
@@ -5,9 +5,9 @@ import classes from "./Navbar.module.css";
 import NavItem from "../Navbar/NavItem";
 import DropdownMenu from "../Navbar/DropdownMenu";
 
-import developmentSvg from "../../Assets/svg/development.svg";
+import developmentSvg from "../../../Assets/svg/development.svg";
 
-import navOptions from "../../Assets/svg/selectDropdown.svg";
+import navOptions from "../../../Assets/svg/selectDropdown.svg";
 
 
 
@@ -15,7 +15,7 @@ const Navbar = () => {
     return (
 			<nav className={classes.navbar}>
 				<ul className={classes.navbarNav}>
-					<NavItem href="/about/this/site" navText="Dev Planner" icon={developmentSvg} />
+					<NavItem href="/myPlan" icon={developmentSvg} />
 					<NavItem icon={navOptions}>
 						<DropdownMenu />
 					</NavItem>
